+++ com.unity.cinemachine/Runtime/Components/CinemachineGroupComposer.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 using Cinemachine.Utility;
 
@@ -12,9 +13,9 @@ namespace Cinemachine
     /// will adjust the FOV and the camera distance to ensure that the entire group of targets
     /// is framed properly.
     /// </summary>
-    [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [AddComponentMenu("")] // Don't display in add component menu
     [SaveDuringPlay]
+    [CameraPipeline(CinemachineCore.Stage.RotationControl)]
     public class CinemachineGroupComposer : CinemachineComposer
     {
         /// <summary>How much of the screen to fill with the bounding box of the targets.</summary>
@@ -23,7 +24,6 @@ namespace Cinemachine
         public float m_GroupFramingSize = 0.8f;
 
         /// <summary>What screen dimensions to consider when framing</summary>
-        [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
         public enum FramingMode
         {
             /// <summary>Consider only the horizontal dimension.  Vertical framing is ignored.</summary>