+++ com.unity.cinemachine/Runtime/Impulse/CinemachineImpulseListener.cs
@@ -11,7 +11,6 @@ namespace Cinemachine
     /// </summary>
     [SaveDuringPlay]
     [AddComponentMenu("")] // Hide in menu
-    [DocumentationSorting(DocumentationSortingAttribute.Level.UserRef)]
     [ExecuteAlways]
     [HelpURL(Documentation.BaseURL + "manual/CinemachineImpulseListener.html")]
     public class CinemachineImpulseListener : CinemachineExtension
@@ -22,7 +21,7 @@ namespace Cinemachine
         /// </summary>
         [Tooltip("When to apply the impulse reaction.  Default is after the Noise stage.  "
             + "Modify this if necessary to influence the ordering of extension effects")]
-        public CinemachineCore.Stage m_ApplyAfter = CinemachineCore.Stage.Aim; // legacy compatibility setting
+        public CinemachineCore.Stage m_ApplyAfter = CinemachineCore.Stage.RotationControl; // legacy compatibility setting
 
         /// <summary>
         /// Impulse events on channels not included in the mask will be ignored.